Revert "move functions to Console.h/cpp"
This reverts commit 67880017774014339d745db258dbabe43bf9963f.
