+++ cpp/open3d/Open3D.h.in
@@ -89,11 +89,10 @@
 #include "open3d/t/pipelines/odometry/RGBDOdometry.h"
 #include "open3d/t/pipelines/registration/Registration.h"
 #include "open3d/t/pipelines/registration/TransformationEstimation.h"
-#include "open3d/utility/Console.h"
+#include "open3d/utility/Logging.h"
 #include "open3d/utility/Eigen.h"
 #include "open3d/utility/FileSystem.h"
 #include "open3d/utility/Helper.h"
-#include "open3d/utility/Logging.h"
 #include "open3d/utility/Timer.h"
 #include "open3d/visualization/gui/Application.h"
 #include "open3d/visualization/gui/Button.h"
