+++ cpp/open3d/core/kernel/ParallelUtil.h
@@ -26,10 +26,6 @@
 
 #pragma once
 
-#ifdef _OPENMP
-#include <omp.h>
-#endif
-
 namespace open3d {
 namespace core {
 namespace kernel {
