+++ cpp/open3d/utility/Logging.h
@@ -26,7 +26,7 @@
 
 #pragma once
 
-#include <functional>
+#include <Eigen/Core>
 #include <iostream>
 #include <string>
 #include <vector>
@@ -301,6 +301,33 @@ private:
 };
 
 std::string GetCurrentTimeStamp();
+
+std::string GetProgramOptionAsString(int argc,
+                                     char **argv,
+                                     const std::string &option,
+                                     const std::string &default_value = "");
+
+int GetProgramOptionAsInt(int argc,
+                          char **argv,
+                          const std::string &option,
+                          const int default_value = 0);
+
+double GetProgramOptionAsDouble(int argc,
+                                char **argv,
+                                const std::string &option,
+                                const double default_value = 0.0);
+
+Eigen::VectorXd GetProgramOptionAsEigenVectorXd(
+        int argc,
+        char **argv,
+        const std::string &option,
+        const Eigen::VectorXd default_value = Eigen::VectorXd::Zero(0));
+
+bool ProgramOptionExists(int argc, char **argv, const std::string &option);
+
+bool ProgramOptionExistsAny(int argc,
+                            char **argv,
+                            const std::vector<std::string> &options);
 
 }  // namespace utility
 }  // namespace open3d
